+++ tests/test_csv_infer.py
@@ -10,29 +10,28 @@ from tests.util import get_basic_sql_reference
 
 
 @pytest.fixture
-def basic_stream_ndjson():
+def basic_stream_csv():
     """
-    A stream of input data. Here, in ndjson format, aka. jsonl.
-
-    http://ndjson.org/
+    A stream of input data. Here, in CSV format.
     """
     return io.StringIO(
         """
-{"id":1,"name":"foo","date":"2014-10-31 09:22:56","fruits":"apple,banana","price":0.42}
-{"id":2,"name":"bar","date":null,"fruits":"pear","price":0.84}
+id,name,date,fruits,price
+1,"foo","2014-10-31T09:22:56","apple,banana",0.42
+2,"bar",\\N,"pear",0.84
     """.lstrip()
     )
 
 
-def test_ndjson_infer_library_success(basic_stream_ndjson):
+def test_csv_infer_library_success(basic_stream_csv):
     """
     Verify basic library use.
     """
     table_name = "foo"
     sg = SchemaGenerator(
         resource=Resource(
-            data=basic_stream_ndjson,
-            content_type="ndjson",
+            data=basic_stream_csv,
+            content_type="csv",
         ),
         target=SqlTarget(
             dialect="crate",
@@ -46,12 +45,12 @@ def test_ndjson_infer_library_success(basic_stream_ndjson):
     assert computed == reference
 
 
-def test_ndjson_infer_cli_file_without_tablename(ndjson_file_basic):
+def test_csv_infer_cli_file_without_tablename(csv_file_basic):
     """
     CLI test: Table name is correctly derived from the input file name or data.
     """
     runner = CliRunner()
-    result = runner.invoke(cli, f"infer-ddl --dialect=crate {ndjson_file_basic}")
+    result = runner.invoke(cli, f"infer-ddl --dialect=crate {csv_file_basic}")
     assert result.exit_code == 0
 
     computed = SqlResult(result.stdout).canonical
@@ -59,14 +58,14 @@ def test_ndjson_infer_cli_file_without_tablename(ndjson_file_basic):
     assert computed == reference
 
 
-def test_ndjson_infer_cli_file_with_tablename(ndjson_file_basic):
+def test_csv_infer_cli_file_with_tablename(csv_file_basic):
     """
     CLI test: Table name takes precedence when obtained from the user.
     """
     table_name = "foo"
 
     runner = CliRunner()
-    result = runner.invoke(cli, f"infer-ddl --dialect=crate --table-name={table_name} {ndjson_file_basic}")
+    result = runner.invoke(cli, f"infer-ddl --dialect=crate --table-name={table_name} {csv_file_basic}")
     assert result.exit_code == 0
 
     computed = SqlResult(result.stdout).canonical
@@ -74,8 +73,8 @@ def test_ndjson_infer_cli_file_with_tablename(ndjson_file_basic):
     assert computed == reference
 
 
-@pytest.mark.parametrize("content_type", ["ndjson", "jsonl", "application/x-ldjson"])
-def test_ndjson_infer_cli_stdin_with_content_type(basic_stream_ndjson, content_type: str):
+@pytest.mark.parametrize("content_type", ["csv", "text/csv"])
+def test_csv_infer_cli_stdin_with_content_type(basic_stream_csv, content_type: str):
     """
     CLI test: Read data from stdin.
     """
@@ -85,7 +84,7 @@ def test_ndjson_infer_cli_stdin_with_content_type(basic_stream_ndjson, content_t
     result = runner.invoke(
         cli,
         args=f"infer-ddl --dialect=crate --table-name={table_name} --content-type={content_type} -",
-        input=basic_stream_ndjson.read(PEEK_BYTES),
+        input=basic_stream_csv.read(PEEK_BYTES),
     )
     assert result.exit_code == 0