+++ CHANGES.rst
@@ -14,6 +14,7 @@ in progress
 - Derive table name from input file name
 - Primary key detection
 - Content type detection using file extension or ``--content-type`` option
+- Add support for CSV input format
 
 
 2023-01-xx 0.0.0